SkinScannerApp/backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12        # multipart form uploads
pydantic-settings>=2.6.0        # .env loading

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0

# ── ML ────────────────────────────────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
grad-cam>=1.5.0

# ── Logging ───────────────────────────────────────────────────────────────────
loguru>=0.7.0
